Basic functionality of the game has been added.
User can now choose their player (the "hero") and the character they're competing against (the "enemy"). When the user makes this selection, these characters are moved to their respective spots on the page. There's an "attack" button that runs the attack() function for those two players. If the enemy dies, they are removed from the page. The user must then select another "enemy" from the remaining characters to play against. If the "hero" dies, the game is over.